Create ProtoReflectionService for the v1alpha proto by producing a ServerServiceDefinition constructed from that of the v1 service but with the service and method names and proto descriptors modified.
Issue #6724.

Squashed commit of the following:

commit f1f2da7be8c1636ca4d668d0eebea54b4ec76abb
Author: Kannan J <kannanjgithub@google.com>
Date:   Tue May 28 10:24:32 2024 +0000

    Some nitty changes.

commit 34f81e56ab973b3c3a20989d960744b5d75e41a1
Author: Kannan J <kannanjgithub@google.com>
Date:   Tue May 28 10:09:49 2024 +0000

    Create ProtoReflectionService for the v1alpha proto by producing a ServerServiceDefinition constructed from that of the v1 service but with the service and method names and proto descriptors modified.

commit e99a8da0900dc093dfc0bed29b19cfa607b42d0a
Author: Kannan J <kannanjgithub@google.com>
Date:   Fri May 24 11:20:28 2024 +0000

    Create V1 alpha service wrapping underlying V1 service, by modifying the ServerServiceDefinition.

commit b2bb9e11a2650e0b21fc5f0c98a4670c3c7acbfc
Author: Kannan J <kannanjgithub@google.com>
Date:   Fri May 24 07:14:42 2024 +0000

    Changes to allow unit testing.

commit a002bc24a055b8301f179615e76307018c798bab
Author: Kannan J <kannanjgithub@google.com>
Date:   Wed May 22 11:53:27 2024 +0000

    Issue #6724 to create a V1 reflection service.
    Changes still in progress on the proto field mapping.
